Move generalized note encryption functionality to zcash_note_encryption crate.
