+++ components/zcash_note_encryption/Cargo.toml
@@ -11,3 +11,12 @@ license = "MIT OR Apache-2.0"
 edition = "2018"
 
 [dependencies]
+blake2b_simd = "0.5"
+byteorder = "1"
+crypto_api_chachapoly = "0.4"
+ff = "0.8"
+group = "0.8"
+rand_core = "0.5.1"
+
+[dev-dependencies]
+zcash_primitives = { version = "0.4", path = "../../zcash_primitives" }
