+++ zcash_primitives/Cargo.toml
@@ -37,6 +37,7 @@ ripemd160 = { version = "0.9", optional = true }
 secp256k1 = { version = "0.20", optional = true }
 sha2 = "0.9"
 subtle = "2.2.3"
+zcash_note_encryption = { version = "0.0", path = "../components/zcash_note_encryption" }
 
 # Temporary workaround for https://github.com/myrrlyn/funty/issues/3
 funty = "=1.1.0"
